Changes input to use a directory
Input now should be a directory with the folders of the landsat images as they are downloaded.
The input file defines the region to consider when performing the textures.
As a result, an output files gets created in each image directory. This file's name is created with the common part of the string that define the landsat 8 images, plus a "text.tif" at the end. This files holds one band for each texture.