Changes testing and prediction of images
Testing and prediction of an image can be done at the same time, or not. This is reflected on the UI as well.
If the testing matrix is empty, testing of and image will be done with 25% of the training data.
predict_an_image method does not fit the random forest anymore. That is only done in the fit_and_calculate_metrics method.